requirements.txt
affine==2.4.0
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
imageio==2.25.0
joblib==1.2.0
networkx==3.0
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
PyWavelets==1.4.1
PyYAML==6.0
rasterio==1.3.5.post1
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
snuggs==1.4.7
threadpoolctl==3.1.0
tifffile==2023.2.3
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.64.1
typing_extensions==4.4.0
